% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiral.R
\name{spiral}
\alias{spiral}
\title{Spiral plot}
\usage{
spiral(
  x,
  n = 5,
  r = 25,
  m = -1,
  bg_fill = "white",
  line_col = "black",
  line_width = 0.05,
  buffer = 0.9,
  infill = NA,
  ...
)
}
\arguments{
\item{x}{numeric vector to plot in a spiral}

\item{n}{number of 'rings' in spiral (default = 5)}

\item{r}{centre hole size as percentage of spiral thickness (default = 25)}

\item{m}{direction of spiral (-1 spiral in towards centre, 1 spiral away from centre (default = -1))}

\item{bg_fill}{background fill colour}

\item{line_col}{spiral line colour}

\item{line_width}{spiral line width}

\item{buffer}{spacing buffer (default = 0.9)}

\item{infill}{in-fill value for extra points needed to fill final spiral ring (default = NA)}

\item{...}{further arguments passed to \code{ggplot2::geom_line()}}
}
\value{
A ggplot2 plot
}
\description{
This code is designed for artistic/aesthetic visualistion of numbers only. It should not be used for serious data visualisation for the following two reasons
 \itemize{
   \item The data may not be fully or correctly reproduced on the spiral
   \item Polar coordinates have major perceptual problems
  }
}
\details{
The data is scaled to sit symmetrically above and below the spiral line
}
\examples{
spiral(rnorm(5000), n=8)
}
